[Test] Correct the BlockDnsQuery_FlaggedOff test
Test clears the wrong fake private DNS front-end. It was passing
previously because the test didn't flush DNS cache. On the second
iteration of the loop, the DNS resolver does not send out the query. It
returns results from the cache provided by the first iteration.

The correct behavior should be clearing the right private DNS front-end
and clearing DNS resolver's cache after each iteration of for-loop.

Bug: N/A
Test: atest resolv_integration_test
Change-Id: I0f87c5cb36bf1887339a3c48bee37358dfac5ae2