Make resolv_set_nameservers take the AIDL parcel directly.
Currently, the code unparcels the parcel and passes its elements
to resolv_set_nameservers as function parameters. These are then
copied to other data structures.

Instead of doing this, just pass the parcel around directly. It
simplifies the code, and it makes it much easier to extend the
parcel in the future.

Test: m
Change-Id: I924b05980dae5c8a343f612f09dbd83d6f01e287